Created views module (#299)
* Created views module

* Fixed imports
